+++ InformationProtection/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## 0.9.2 (2025-09-12)
+
++ Upd: Label Object - Renamed "FQLA" property to "FQLN"
++ Fix: Log Path - Changed to LocalAppData from current path
+
 ## 0.9.0 (2025-09-12)
 
 + Initial Release